refactor: extract overlay focus logic to separate mixin
